requirements.txt
agentops==0.3.12
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
boto3==1.35.26
botocore==1.35.26
build==1.2.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
cleo==2.1.0
click==8.1.7
cohere==5.9.4
coloredlogs==15.0.1
crashtest==0.4.1
crewai==0.63.6
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
docstring_parser==0.16
dulwich==0.21.7
durationpy==0.7
embedchain==0.1.122
fastapi==0.115.0
fastavro==1.9.7
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-aiplatform==1.68.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
installer==0.7.0
instructor==1.3.3
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kubernetes==31.0.0
langchain==0.2.16
langchain-cohere==0.1.9
langchain-community==0.2.17
langchain-core==0.2.41
langchain-experimental==0.0.65
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.128
litellm==1.48.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mem0ai==0.1.16
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
neo4j==5.24.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.48.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parameterized==0.9.0
pexpect==4.9.0
pkginfo==1.11.1
platformdirs==4.3.6
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
portalocker==2.10.1
posthog==3.6.6
proto-plus==1.24.0
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
qdrant-client==1.11.3
rank-bm25==0.2.2
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
schema==0.7.7
SecretStorage==3.3.3
setuptools==72.1.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.20.0
tomlkit==0.13.2
tqdm==4.66.5
trove-classifiers==2024.9.12
typer==0.12.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
virtualenv==20.26.5
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wheel==0.44.0
wrapt==1.16.0
yarl==1.12.1
zipp==3.20.2
